requirements.txt
numpy==1.11.2
Pillow==3.4.2
pkg-resources==0.0.0
protobuf==3.1.0
scipy==0.18.1
six==1.10.0
tensorflow-gpu==0.12.0rc0
